+++ Domain/Repositories/IOrderRepository.cs
@@ -5,9 +5,9 @@ namespace DDDSample.Domain.Repositories;
 public interface IOrderRepository
 {
     Task<IEnumerable<Order>> GetAllAsync();
-    Task<Order?> GetByIdAsync(int id);
+    Task<Order?> GetByIdAsync(long id);
     Task AddAsync(Order order);
     Task UpdateAsync(Order order);
     Task DeleteAsync(Order order);
-    Task<bool> ExistsAsync(int id);
+    Task<bool> ExistsAsync(long id);
 }
